Added id_map try 7

+++ preprocess/preprocess_sun2cao.sh
@@ -14,24 +14,27 @@ do
 done
 }
 
-if true
-then
-    #[train.dat, valid.dat, test.dat by splitting rating-delete-missing-item.txt]
-    python sun2cao_split.py --loadfile '../../datasets/ml1m-sun/ml1m/rating-delete-missing-itemid.txt' --column 'user_id' --umapfile '../../datasets/ml1m-sun2cao/ml1m/u_map.dat' --imapfile '../../datasets/ml1m-sun2cao/ml1m/i_map.dat' --savepath '../../datasets/ml1m-sun2cao/ml1m/' &
-    BACK_PID=$!
-    wait $BACK_PID
-else
-    #[train.dat, valid.dat, test.dat symbolic links]
-    ln -s ~/git/datasets/ml1m-cao/ml1m/train.dat ~/git/datasets/ml1m-sun2cao/ml1m/train.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/valid.dat ~/git/datasets/ml1m-sun2cao/ml1m/valid.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/test.dat ~/git/datasets/ml1m-sun2cao/ml1m/test.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/i_map.dat ~/git/datasets/ml1m-sun2cao/ml1m/i_map.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/u_map.dat ~/git/datasets/ml1m-sun2cao/ml1m/u_map.dat
-fi
-
 #[activate jointrec]
 conda deactivate
 conda activate jointrec
+
+if no_exist "~/git/datasets/ml1m-sun2cao/ml1m/train.dat"
+then
+    if true
+    then
+        #[train.dat, valid.dat, test.dat by splitting rating-delete-missing-item.txt]
+        python sun2cao_split.py --loadfile '../../datasets/ml1m-sun/ml1m/rating-delete-missing-itemid.txt' --column 'user_id' --umapfile '../../datasets/ml1m-sun2cao/ml1m/u_map.dat' --imapfile '../../datasets/ml1m-sun2cao/ml1m/i_map.dat' --savepath '../../datasets/ml1m-sun2cao/ml1m/' &
+        BACK_PID=$!
+        wait $BACK_PID
+    else
+        #[train.dat, valid.dat, test.dat symbolic links]
+        ln -s ~/git/datasets/ml1m-cao/ml1m/train.dat ~/git/datasets/ml1m-sun2cao/ml1m/train.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/valid.dat ~/git/datasets/ml1m-sun2cao/ml1m/valid.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/test.dat ~/git/datasets/ml1m-sun2cao/ml1m/test.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/i_map.dat ~/git/datasets/ml1m-sun2cao/ml1m/i_map.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/u_map.dat ~/git/datasets/ml1m-sun2cao/ml1m/u_map.dat
+    fi
+fi
 
 #[clean_auxiliary.txt]
 if no_exist "../../datasets/ml1m-sun2cao/ml1m/clean_auxiliary.txt"
