+++ summarization/summarize_sun-genre.sh
@@ -50,19 +50,22 @@ conda activate jointrec
 
 cd ../preprocess
 
-if true
+if no_exist "~/git/datasets/ml1m-sun2cao/ml1m/train.dat"
 then
-    #[train.dat, valid.dat, test.dat by splitting rating-delete-missing-item.txt]
-    python sun2cao_split.py --loadfile '../../datasets/ml1m-sun/ml1m/rating-delete-missing-itemid.txt' --column 'user_id' --umapfile '../../datasets/ml1m-summarized_sun/ml1m/u_map.dat' --imapfile '../../datasets/ml1m-summarized_sun/ml1m/i_map.dat' --savepath '../../datasets/ml1m-summarized_sun/ml1m/' &
-    BACK_PID=$!
-    wait $BACK_PID
-else
-    #[train.dat, valid.dat, test.dat symbolic links]
-    ln -s ~/git/datasets/ml1m-cao/ml1m/train.dat ~/git/datasets/ml1m-summarized_sun/ml1m/train.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/valid.dat ~/git/datasets/ml1m-summarized_sun/ml1m/valid.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/test.dat ~/git/datasets/ml1m-summarized_sun/ml1m/test.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/i_map.dat ~/git/datasets/ml1m-summarized_sun/ml1m/i_map.dat
-    ln -s ~/git/datasets/ml1m-cao/ml1m/u_map.dat ~/git/datasets/ml1m-summarized_sun/ml1m/u_map.dat
+    if true
+    then
+        #[train.dat, valid.dat, test.dat by splitting rating-delete-missing-item.txt]
+        python sun2cao_split.py --loadfile '../../datasets/ml1m-sun/ml1m/rating-delete-missing-itemid.txt' --column 'user_id' --umapfile '../../datasets/ml1m-summarized_sun/ml1m/u_map.dat' --imapfile '../../datasets/ml1m-summarized_sun/ml1m/i_map.dat' --savepath '../../datasets/ml1m-summarized_sun/ml1m/' &
+        BACK_PID=$!
+        wait $BACK_PID
+    else
+        #[train.dat, valid.dat, test.dat symbolic links]
+        ln -s ~/git/datasets/ml1m-cao/ml1m/train.dat ~/git/datasets/ml1m-summarized_sun/ml1m/train.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/valid.dat ~/git/datasets/ml1m-summarized_sun/ml1m/valid.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/test.dat ~/git/datasets/ml1m-summarized_sun/ml1m/test.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/i_map.dat ~/git/datasets/ml1m-summarized_sun/ml1m/i_map.dat
+        ln -s ~/git/datasets/ml1m-cao/ml1m/u_map.dat ~/git/datasets/ml1m-summarized_sun/ml1m/u_map.dat
+    fi
 fi
 
 #[sum_auxiliary.txt]
